backend/requirements.txt
chromadb==1.0.21
faiss_cpu==1.12.0
fastapi==0.116.1
langchain==0.3.27
langchain_community==0.3.29
langchain-core==0.3.76
langchain_text_splitters==0.3.11
openai==1.101.0
pydantic==2.11.2
python-dotenv==1.0.1
SQLAlchemy==2.0.43
pdfplumber==0.11.7
PyMySQL==1.1.2
python-docx==1.2.0
torch==2.3.0
transformers==4.56.0
unstructured==0.18.15